requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
fastapi==0.101.0
groq==0.31.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
jiter==0.10.0
openai==1.100.2
pydantic==2.11.7
pydantic_core==2.33.2
python-telegram-bot==20.4
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.23.2
python-dotenv
